+++ .gitignore
@@ -10,13 +10,6 @@ Thumbs.db
 node_modules/
 .vscode/
 .github/
-.babelrc
-.gitignore
-.prettierrc.json
 release-notes.md
-webpack.common.js
-webpack.dev.js
-webpack.prod.js
-package.json
-package-lock.json
+
