+++ package.json
@@ -1,0 +1,37 @@
+{
+  "name": "demo-theme",
+  "version": "1.0.0",
+  "description": "Dawn theme used to build drawer cart using react",
+  "main": "index.js",
+  "scripts": {
+    "dev": "webpack --config webpack.dev.js --progress --color",
+    "build": "webpack --config webpack.prod.js --progress --color",
+    "shopify": "shopify theme dev --store=my-demo-store-01.myshopify.com",
+    "format": "prettier --write './src/**/*.{js,jsx,css}' --config ./.prettierrc"
+  },
+  "author": "Isaac Awonugba",
+  "license": "ISC",
+  "dependencies": {
+    "currency-symbol-map": "^5.1.0",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-icons": "^4.11.0"
+  },
+  "devDependencies": {
+    "@babel/core": "^7.23.2",
+    "@babel/preset-env": "^7.23.2",
+    "@babel/preset-react": "^7.22.15",
+    "babel-loader": "^9.1.3",
+    "core-js": "^3.33.2",
+    "css-loader": "^6.8.1",
+    "eslint": "^8.53.0",
+    "eslint-plugin-prettier": "^5.0.1",
+    "eslint-plugin-react": "^7.33.2",
+    "prettier": "^3.0.3",
+    "sass-loader": "^13.3.2",
+    "style-loader": "^3.3.3",
+    "webpack": "^5.89.0",
+    "webpack-cli": "^5.1.4",
+    "webpack-merge": "^5.10.0"
+  }
+}
