--- src/components/DrawerCart.module.css
@@ -1,28 +1,0 @@
-.reactCart {
-  width: 48rem;
-  position: fixed;
-  right: 0;
-  top: 0;
-  z-index: 9999999;
-  background-color: rgba(255, 255, 255);
-  height: 100vh;
-  padding: 2rem;
-  display: flex;
-  flex-direction: column;
-}
-
-.reactCartActive {
-  transition: ease-in 0.5s;
-  right: 0rem;
-}
-
-.reactCartInactive {
-  right: -48rem;
-  transition: ease-out 0.5s;
-}
-
-.reactCartBody {
-  display: flex;
-  flex-direction: column;
-  gap: 2.5rem;
-}
