+++ webpack.common.js
@@ -1,0 +1,23 @@
+const path = require('path');
+module.exports = {
+  entry: {
+    'cart-drawer': './src/cart-drawer.jsx'
+  },
+  module: {
+    rules: [
+      {
+        test: /\.(js|jsx)$/,
+        exclude: /node_modules/,
+        use: ['babel-loader']
+      },
+      {
+        test: /\.css$/,
+        use: ['style-loader', 'css-loader']
+      }
+    ]
+  },
+  output: {
+    filename: '[name].bundle.js',
+    path: path.resolve(__dirname, 'assets')
+  }
+};
